Squashed commit of the following:
commit 01a469aefa57ccde8bc756ebdbf15484689e8a01
Author: Tibet Tornaci <oofdere@hey.com>
Date:   Fri Jan 28 18:25:37 2022 -0800

    factor out image from app.py

commit aa5a2245f103cc319d784639b6d58b445f976b65
Author: Tibet Tornaci <oofdere@hey.com>
Date:   Fri Jan 28 17:48:59 2022 -0800

    custom error pages

    doesn't work in debug :(
