+++ templates/detail_view.html
@@ -8,6 +8,6 @@
         </nav>
     </aside>
     <div role="document">
-        <img src="images/{{ image.hash }}.{{ image.extension }}" alt="view_image">
+        <img src="/images/{{ image.hash }}.{{ image.extension }}" alt="view_image">
     </div>
 {% endblock %}
