+++ templates/gallery_partial.html
@@ -1,7 +1,7 @@
 <div class="gallery">
     {% for image in images %}
     <div class="gallery-item">
-        <a href="{{ url_for('view', image_hash=image.hash) }}" data-tooltip="{{ image.caption }}">
+        <a href="{{ url_for('image.view', image_hash=image.hash) }}" data-tooltip="{{ image.caption }}">
             <img src="{{ url_for('get_thumb', image_hash=image.hash, filename=image.hash+'.'+image.extension) }}" alt="{{ image.caption }}" />
         </a>
     </div>
